requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-environ==0.11.2
requests==2.32.3
psycopg2-binary>=2.9.10        # PostgreSQL driver – Python 3.12 compatible